+++ src/MapperServiceProvider.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Miniyus\Mapper\Provider;
+namespace Miniyus\Mapper;
 
+use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
-use Miniyus\Mapper\Mapper;
-use Miniyus\Mapper\MapperInterface;
+
 
 class MapperServiceProvider extends ServiceProvider
 {
@@ -15,7 +15,15 @@ class MapperServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->app->bind(MapperInterface::class, Mapper::class);
+        $this->app->bind(MapperInterface::class, function (Application $app) {
+            $configRepository = $app->get('config');
+
+            $mapperConfig = $configRepository->get('mapper');
+
+            return Mapper::newInstance($mapperConfig);
+        });
+
+        $this->app->alias(MapperInterface::class, 'mapper');
     }
 
     /**
@@ -38,9 +46,9 @@ class MapperServiceProvider extends ServiceProvider
         }
 
         $this->publishes([
-            __DIR__ . '/../Stubs/Map.stub' => base_path('stubs/Map.stub')
+            __DIR__ . '/stubs/Map.stub' => base_path('stubs/Map.stub')
         ]);
 
-        $this->commands(\Miniyus\Mapper\Commands\GenerateMap::class);
+        $this->commands(\Miniyus\Mapper\Consonle\GenerateMap::class);
     }
 }