+++ tests/Data/DemoDto.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Data;
+
+use Miniyus\Mapper\Data\Dto;
+
+class DemoDto extends Dto
+{
+
+}