+++ tests/Data/DemoEntity.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Data;
+
+class DemoEntity extends \Miniyus\Mapper\Data\Entity
+{
+
+    /**
+     * @inheritDoc
+     */
+    protected function getIdentifier(): string
+    {
+        return 'demo';
+    }
+}