+++ tests/Data/DemoMap.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Data;
+
+use Miniyus\Mapper\Data\Dto;
+use Miniyus\Mapper\Data\Entity;
+
+class DemoMap extends \Miniyus\Mapper\Maps\Map
+{
+
+    /**
+     * @inheritDoc
+     */
+    protected function toEntity(Dto $dto, Entity $entity)
+    {
+        // TODO: Implement toEntity() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    protected function toDto(Entity $entity, Dto $dto)
+    {
+        // TODO: Implement toDto() method.
+    }
+}