+++ tests/MapperTest.php
@@ -1,0 +1,5 @@
+<?php
+
+class MapperTest extends \PHPUnit\Framework\TestCase
+{
+}